Add error field to Result and move to internal
